requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
dora_search==0.1.12
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.18.0
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
idna==3.10
ImageHash==4.3.1
Jinja2==3.1.6
joblib==1.4.2
julius==0.2.7
keras==3.7.0
kiwisolver==1.4.7
lameenc==1.8.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.2
numpy==2.0.2
omegaconf==2.3.0
opencv-python==4.10.0.84
openpyxl==3.1.5
openunmix==1.3.0
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==5.29.0
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==80.4.0
six==1.16.0
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
submitit==1.5.2
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
treetable==0.2.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
